Time: 3 ms (56.33%), Space: 6 MB (47.88%) - LeetHub

+++ 191-number-of-1-bits/191-number-of-1-bits.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int hammingWeight(uint32_t n) {
         uint8_t numOnes = 0;
-        for(int i = 0; i < 32; i++){
+        for(uint8_t i = 0; i < 32; i++){
             numOnes += n & 1;
             n >>= 1;
